hasRole checks for object and id
We can now check if an user has a role by passing an object or the id of the role as well! Instead of only by name. Also extracted a method for multiple roles checks so we can check by both the hasRole method for multiple roles as well as by calling the new hasRoles method. Keeps the code a little cleaner!