call to auth.model in boot()
On the last version, like in EntrustRoleTrait, there is a switch to Config::get('auth.providers.users.model') which is better than  Config::get('auth.model') that was not set, even following the install documentation. I looked for all the old calls in my code and found this call in dev-master. I think this is a mistake. So ... is it ?